Make WakeLockRequestOptions' signal non-nullable, do not set a default
In Web IDL, the following `WakeLockRequestOptions` instances in JavaScript
should have the same effect (a dictionary that does _not_ have the `signal`
member present):

    {}
    {signal: undefined}

which are different from

    {signal: null}

which means a `WakeLockRequestOptions` dictionary with one member, `signal`,
whose value is `null`.

Setting a default value to the nullable member `signal` means we'd always
have a `signal` member in the dictionary even when we do not set it at
all (first 2 cases), which is not needed and does not make sense.

Additionally, `signal` does not need to be nullable to begin with: by
default, dictionary members do not need to be present, so just make it a
regular, non-nullable member that can either not be present or be set to an
AbortSignal, but not null.