Adjust the "Release Wake Lock" algorithm after #209
That commit removed all the explicit checks for AbortSignal's aborted flag,
and it relies on "release wake lock" to clean up if the aborted flag is set
between any of the parallel steps in WakeLock.request().

Make the "release wake lock" algorithm abort earlier if a given promise is
not in [[ActiveLocks]], which can happen if AbortController.abort() is
called after querying for permission to get a wake lock but before actually
acquiring it.