Explain why WakeLock.request() has no "if aborted" step.
Algorithms with an "abort when" part usually are followed by an "if aborted"
step describing what to do if a series of steps was aborted.

Explain why we do not have one here: the only way for the "abort when"
condition to be reached is if one calls `AbortController.abort()`, in which
case we call "release wake lock" and that rejects the promise and does
everything we would do in an "if aborted" step.