Add another condition to WakeLock.request()'s "abort when" block.
In the WakeLock.request() steps that run in parallel, there was a short
window of time before reaching the "Let success be the result of awaiting
acquire a wake lock with promise and type" substep where `promise` is not
yet in `[[ActiveLocks]]`.

In practice, this means a change in document visibility or loss of document
activity would not cause it to be rejected.

For now, only changes in document visibility really have an effect here, so
extend the "abort when" clause with an extra check.

Fixes #222.