Remove the note on page visibility handling from WakeLock.request()
This note was added back in commit 16cd4c749b4 ("Spec update"), but commit
34119d275c ("Handle document visibility") changed the spec in a way that
contradicts the note: we no longer handle visibility changes automatically,
and leave it up to script authors to do that instead.

Whether this is a good idea or at least be made optional is still being
discussed; in the meantime, remove the note since it does not reflect
reality.

Related to #139