Add an "if aborted" step to WakeLock.request() to deal with hidden documents
This is a follow-up to commit 37ad6b0c ("Add another condition to
WakeLock.request()'s "abort when" block)" that attempted to fix #222.

After #236 we no longer have an `options` argument, but we still need to do
something with `|promise|` if we hit the "abort when" check in the
algorithm.

Add an "if aborted" step that takes care of it, and rejects `|promise|` with
a `NotAllowedError`, the same error that is thrown when a document is
already hidden earlier in `WakeLock.requests()`'s algorithm.

Fixes #239.