Rewrite user-visible API.
Address #226 by redefining the way script authors are supposed to interact
with the Wake Lock API.

The biggest issue with the previous iteration of the API is that it was
_too_ simple and ended up using the promise returned by `WakeLock.request()`
as a way to indicate whether a lock was being held or not. This is not
intuitive at all, as the promise would never resolve while the lock wasn't
released.

The new API is a compromise between the 2018 version of the API and the one
we had until now:

* We go back to adding a partial interface to `Navigator`, but we add a
  partial interface to `WorkerNavigator` as well.

* The `WakeLock` interface's only remaining method is no longer static, and
  instead of returning a `Promise<void>` it now returns a
  `Promise<WakeLockSentinel>`.

* `WakeLockSentinel` is a new interface that represents a handle to a
  platform wake lock returned by `WakeLock.request()`. It has a `release()`
  method and an `onrelease` event handler. `release()` replaces the previous
  `AbortController`-based infrastructure for releasing a lock, and
  `onrelease` lets users be notified when a lock is released (since they can
  also be released by the UA under circumstances such as loss of full
  activity).
  When all `WakeLockSentinel`s of a given type are released, the platform
  lock is released too.

* `WakeLockEvent` and its related interface `WakeLockEventInit` are
  delivered to `WakeLockSentinel`'s `onrelease` event handler, and allow
  users to also re-request a lock if necessary.