Remove WakeLockEvent and WakeLockEventInit.
Per https://w3ctag.github.io/design-principles/#state-and-subclassing, we do
not really need a custom event class here.

`WakeLockEvent` has a single attribute, `lock`, that points to the event
target. That's exactly what `Event.target` also does, so we can just create
a simple `Event` instance and let the "fire an event" DOM algorithm set
`target` for us to achieve the same thing. In fact, it already does, so
before this commit we had

    ev.lock === ev.target

Fixes #238.