Make WakeLockPermissionDescriptor.type required.
Different wake lock types have different permission models, so the previous
assumption that not specifying `type` would ask for permission for all wake
lock types does not hold.

The "block a permission" and "obtain permission" algorithms did not have to
be updated, as they were already assuming `type` was valid.

Closes #233.