Stop checking if "acquire the wake lock" succeeds.
Not to be confused with the "acquire a wake lock" algorithm. The latter used
to call the former and check if it succeeded, which indicated that a lock
had been successfully obtained from the operating system.

This could potentially leak user information, as an "acquire the wake lock"
failure could indicate that the user's system was in low power mode or had a
low battery level.

Now failures to acquire an operating system lock are indistiguishable from
successful acquisitions; in both cases, WakeLock.request() will ultimately
resolve the promise it returns with a WakeLockSentinel.

This spec change also matches the current Blink implementation, which never
checked if the platform wake lock acquisition succeeded. This implementation
has been shipping to users with no adverse feedback.

Fixes #247.