editorial: Remove the "obtain a permission" algorithm.
This may have made more sense back when we had a custom
`WakeLockPermissionDescriptor`, but even then this felt like implementing a
custom permission model for no reason.

We are just interested in knowing whether we have permission to request a
wake lock from the platform, and we can use the Permission spec's "request
permission to use" algorithm for that just fine.

One important change is related to checking transient activation: the
"obtain a permission" algorithm essentially denied requests when
"screen-wake-lock"'s permission state was set to "prompt" AND there was no
transient user activation. In general, an API should either always or never
be user-activation gated. Given Blink is the only current implementation of
this spec and it does not currently gate this API on user activation, the
transient user activation check was removed altogether for now.

Together with #297, this fixes #187 and fixes #198.