editorial: Formally define a permission revocation algorithm. (#297)
The "block a permission" algorithm was manipulating a `PermissionDescriptor`
instance to change its "permission state" and then release any existing
locks. However, the Permissions spec does not actually define a way for
other specifications to modify a PermissionDescriptor's permission state.

What we want here is to react to a user-triggered change that sets the
permission state to "denied". That's what the "permission revocation
algorithm" hook in the Permissions spec does, so instead of defining a
"block a permission" algorithm define a custom "permission revocation
algorithm" instead.

Related to #198 -- one of @marcoscaceres' comments is about hooking more of
the permissions text directly into the Permissions spec.