editorial: Improve references to terms from the Page Visibility spec. (#306)
* @annevk says in #299 that, as a general rule, invoking IDL members from
  algorithms, like we were doing with `document.hidden`, is not recommended.
  Instead, invoke the "determine the visibility state" steps defined in the
  Page Visibility spec directly.

* Clean up the "Dependencies" section; for now we only need one term that is
  not exported by the latest Page Visibility TR (this is fixed in the ED):
  the "now hidden algorithm". The steps described in the "Handling document
  loss of visibility" steps now explicitly run after that algorithm, which
  runs when the top-level browsing context's Document becomes hidden.

More generally, these changes should also make it more clear that the
visibility state checks only apply to the top-level browsing
context (according to the Page Visibility spec itself). Although we were
previously checking if a given `Document` was `hidden` even if it did not
belong to a top-level browsing context, that was simply invalid.