editorial: Remove the "state record" concept.
This was originally discussed in #311. The "platform wake lock" and "state
record" concepts were used in a confusing way across the spec.

The idea behind "state record" and the `[[ActiveLocks]]` internal slot was
simply to represent a collection of `WakeLockSentinel` instances associated
with a given `Document`. We can thus be more explicit and add a new
`[[ActiveLocks]]` internal slot to the `Document` interface, which is now a
mapping of wake lock types -> WakeLockSentinel instances (the mapping
existed before, but was defined in prose).

This change should have no user-visible consequences.