Revert "workflows/tidy: Run on ubuntu-20.04. (#317)"
This reverts commit b6b47420748ee0975b6be4a919a28aaa4df1ae4e. The tidy 5.6.0
package shipped in Debian (up to bullseye at the time of writing) as well as
Ubuntu (up to 21.04 at least) is built with a patch that causes the
configurations settings we pass via `-config .tidyrc` to be ignored when
running (0011-Revert-Remove-unnecessary-AdjustConfig-logic.patch).

The easiest fix is to just revert the commit and continue running tidy on
macos-latest for the time being.