editorial: Reference the "external now hidden" algorithm from Page Visibility
A new Page Visibility TR has been published, and after a few follow-up
commits to the spec it is now possible to reference its "external now
hidden" algorithm, which is invoked when a page is hidden.

This change should have no user-visibile effects, as our current algorithm
was already supposed to run after the "visibilitychange" event is fired and
the page becomes hidden.